Let user type characters in terminal and display.
Define USART_FLAG_RXNE as 0x0020 and use it to detect if there is a serial of
characters typed in terminal by user or not. In this version, I have not
completed the function of "Enter" key. If I press "Enter", cursor will go back
to the first character which I key in before in the same line.